Update `build.gradle` in server and client
We needed to (re)add some distribution config to `server/build.gradle`
so that Gradle would bundle up `index.html` and the other resources
and put them in a place that the Java Spark server could find and
serve from.

This adds the necessary `build` task to `client/build.gradle` that
will build the client in production mode on the droplet, and sets
the URL for the droplet in the production environment.